Extract when package from pipelines
