+++ pkg/when/changein_test.go
@@ -1,4 +1,4 @@
-package pipelines
+package when
 
 import (
 	"testing"
